+++ STRR/src/game/__init__.py
@@ -15,4 +15,4 @@ License: MIT
 
 from typing import Final
 
-__version__: Final[str] = "0.0.1"
+__version__: Final[str] = "0.0.10"